requirements.txt
Django==5.2.7
django-cml2
lxml
pytest
pytest-django
django-oscar
sorl-thumbnail
pycountry
